# src/main/java/com/tamanna/demo/model/dto/CandidateDTO.java
package com.tamanna.demo.model.dto;

public class CandidateDTO {

    private Integer id;
    private String name;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}