# src/main/java/com/tamanna/demo/model/dto/InterviewerDTO.java
package com.tamanna.demo.model.dto;

public class InterviewerDTO {

    private Integer id;
    private String name;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}